requirements.txt
groq>=0.11.0
requests>=2.31.0
beautifulsoup4>=4.12.0
streamlit>=1.35.0
duckduckgo-search>=6.0.0
google-api-python-client>=2.100.0
python-dotenv>=1.0.0
